+++ src/cli/commands/commands/index.ts
@@ -1,2 +1,3 @@
 export * from './package';
 export * from './publish';
+export * from './update-version';