+++ src/cli/commands/commands/package/types.ts
@@ -1,0 +1,13 @@
+export interface Options {
+  readonly config: string;
+}
+
+export interface FinalOptions {
+  readonly projectDirectory: string;
+  readonly configPath: string;
+}
+
+export interface PackageConfig {
+  readonly publishDir: string;
+  readonly include: readonly string[];
+}