+++ src/cli/commands/commands/publish/types.ts
@@ -1,0 +1,15 @@
+export interface Options {
+  readonly config: string;
+  readonly dryRun: boolean | undefined;
+}
+
+export interface FinalOptions {
+  readonly projectDirectory: string;
+  readonly configPath: string;
+  readonly dryRun: boolean;
+}
+
+export interface PublishConfig {
+  readonly publishDir: string;
+  readonly dryRun: boolean;
+}