+++ src/cli/program.ts
@@ -1,8 +1,20 @@
 import nodePath from 'node:path';
 import { Command } from 'commander';
 import { readPackageJsonSync } from '@gmjs/package-json';
+import {
+  addCommandPackage,
+  addCommandPublish,
+  addCommandUpdateVersion,
+} from './commands';
 
-export function createProgram(): Command {
+export async function execute(args: readonly string[]): Promise<void> {
+  const program = createProgram();
+  setupProgram(program);
+
+  await program.parseAsync(args);
+}
+
+function createProgram(): Command {
   const program = new Command();
   program
     .name('pnpmpub')
@@ -12,4 +24,10 @@ export function createProgram(): Command {
     );
 
   return program;
+}
+
+function setupProgram(program: Command): void {
+  addCommandPackage(program);
+  addCommandPublish(program);
+  addCommandUpdateVersion(program);
 }