+++ src/run.ts
@@ -1,13 +1,3 @@
-import { addCommandPackage, addCommandPublish, createProgram } from './cli';
-import { addCommandUpdateVersion } from './cli/commands/commands/update-version';
+import { execute } from './cli';
 
-export async function run(): Promise<void> {
-  const program = createProgram();
-  addCommandPackage(program);
-  addCommandPublish(program);
-  addCommandUpdateVersion(program);
-
-  await program.parseAsync(process.argv);
-}
-
-run();
+execute(process.argv);